requirements.txt
streamlit

boto3
botocore

pandas
numpy
requests
python-dotenv

amazon-transcribe

audio-recorder-streamlit==0.0.10

streamlit-webrtc==0.64.5
aiortc==1.14.0

tqdm
pdfplumber
python-docx